refactor: Extract request building logic into RequestBuilder trait
- Move request building methods from Request class to new RequestBuilder trait
- Simplify Request class by utilizing RequestBuilder trait
